\name{optimal_clust}
\alias{optimal_clust}
\title{Cluster validation for cluster identification}
\description{This function validates the clustering resulting from Variability-based Neighbor Clustering as per \code{vnc} by calculating silhouette coefficients and returns the most appropriate clustering of the data as well as potentional outliers.}
\usage{
optimal_clust(x, y)
}
\arguments{
  \item{x}{an object of class \code{hclust} as generated by the \code{vnc} function.}
  \item{y}{the distance matrix which was also used as input for \code{vnc} (an object of class \code{dist}).}
}
\details{
	
The \code{optimal_clust} function identifies the optimal number of clusters, i.e., time stages, after Variability-based Neighbor Clustering (VNC, Gries and Hilpert 2008, 2012) has been conducted via the function \code{vnc} and also identifies potential outliers. The optimal clustering is identified on the basis of the calculation of silhouette values  (Rousseeuw 1987) using R's \code{silhouette} function. 

Silhouette values provide information about the  consistency of clusters by measuring the dissimilarity of an object to the cluster that it is in, compared to its dissimilarity to other clusters. A large silhouette value, i.e., a value close to 1, indicates that the object is clustered well as it is, and a negative value indicates that the object has been assigned to the wrong cluster. The silhouette coefficient of a cluster is moreover defined as the average of silhouettes in a cluster. 

The function \code{optimal_clust} takes an object of class hclust as generated by \code{vnc} and the corresponding distance matrix as calculated via \code{distvnc} as input. \code{optimal_clust} iterates through all clustering possibilities according to the possible number of clusters (i.e., merges) throughout the clustering process, which are accessed via \code{cutree}, and calculates the average of silhouette coefficients of all clusters in a clustering via the \code{silhouette} function. Eventually, \code{optimal_clust} identifies the clustering with the highest average silhouette coefficient as the best candidate, and returns information about the cluster memberships of data points with respect to the optimal clustering. 

\code{optimal_clust} is moreover part of the implementation of the DiaHClust methodology (Schätzle and Booth 2019) in the form of the \code{diahclust} function which provides an iterative approach to VNC in order to arrive at a suitable number of clusters for the identification of stages in language change.

}
\value{
	An object of class \code{optimal_clust}, including a list with the following values:
	\item{opt_clust}{the number of clusters in the clustering with the highest average silhouette coefficient.} 
	\item{final_clust}{the clustering with the highest average silhouette coefficient, provides information about cluster memberships.}
	\item{silcoef}{the average silhouette coefficient of the clustering with the highest average silhouette coefficient.}
}
\references{
Stefan Th. Gries and Martin Hilpert. 2008. The identification of stages in diachronic data: variability-based neighbour clustering. Corpora, 3(1):59–81. 
Stefan Th. Gries and Martin Hilpert. 2012. Variability-based neighbor clustering: A bottom-up approach to periodization in historical linguistics. In Nevalainen Terttu and Elizabeth Closs Traugott, editors, The Oxford Handbook of the History of English, pages 134–144. Oxford University Press, Oxford. 
Rousseeuw, P.J. (1987) Silhouettes: A graphical aid to the interpretation and validation of cluster analysis. J. Comput. Appl. Math., 20, 53–65.

Christin Schätzle and Hannah Booth. 2019. DiaHClust: an iterative hierarchical clustering apprach for identifying stages in language change. to appear.
}
\author{Christin Schätzle}
\seealso{
	\code{\link{silhouette}}
	\code{\link{diahclust}}
	\code{\link{distvnc}}
	\code{\link{vnc}}
}
\examples{
icelandic=data(icelandic)
icelandic.cor=cor(icelandic[,-1])  #[,-1] because rows are labeled
icelandic.dist=dist(icelandic.cor)

icelandic.vnc=vnc(icelandic.dist, method="average")
plot(icelandic.vnc, hang=-1) #plotting the resulting dendrogram

#cluster validation for identification of time stages
optimal=optimal_clust(icelandic.vnc, icelandic.dist) 
}

